Some logging improvements

+++ cmd/repository/dtd_sql_test.go
@@ -375,7 +375,7 @@ func TestDtdRepositorySql_FindFaresForFlow(t *testing.T) {
 		db *gorm.DB
 	}
 	type args struct {
-		flowId uint
+		flowId uint64
 	}
 	tests := []struct {
 		name      string
